# [SPARK-41669][SQL] Early pruning in canCollapseExpressions